refactor: move controller and dto per review feedback
