+++ src/main/java/swyp12/team9/server/api/dto/SimilarContentResponse.java
@@ -1,4 +1,4 @@
-package swyp12.team9.server.domain.recommendation.dto;
+package swyp12.team9.server.api.dto;
 
 public record SimilarContentResponse(
     RecommendationResponse content,
